make 64-bit integers use "long long" on both 64/32 bit, clean up related headers
- remove related _LP64 checks from include/sys/int_*
- use 'j' format specifiers for intmax_t
- sys/int_limits.h: simplify & replace decimal magic numbers with hex
- remove extern "C" from these headers; they have no function decls
- sys/types.h: use standard integer types instead of u_longlong_t etc.
- greg_t: typedef to intptr_t

these cause a number of invalid format specifiers throughout the tree
since we are essentially changing the expected specifier for 64-bit
numbers from "l" to "ll". the payoff is that it will be "ll" regardless
of bit width.

+++ arch/x86/kernel/os/trap.c
@@ -1692,8 +1692,8 @@ dumpregs(struct regs *rp)
 	printf(fmt, "r10", rp->r_r10, "r11", rp->r_r11, "r12", rp->r_r12);
 	printf(fmt, "r13", rp->r_r13, "r14", rp->r_r14, "r15", rp->r_r15);
 
-	printf(fmt, "fsb", rdmsr(MSR_AMD_FSBASE), "gsb", rdmsr(MSR_AMD_GSBASE),
-	    " ds", rp->r_ds);
+	printf(fmt, "fsb", (uintptr_t)rdmsr(MSR_AMD_FSBASE), "gsb",
+	    (uintptr_t)rdmsr(MSR_AMD_GSBASE), " ds", rp->r_ds);
 	printf(fmt, " es", rp->r_es, " fs", rp->r_fs, " gs", rp->r_gs);
 
 	printf(fmt, "trp", rp->r_trapno, "err", rp->r_err, "rip", rp->r_rip);
